# Speed up parse_header of the binary IndexData