The change makes a measurable performance gain for debuginfod-rs:

```
Before:
  Time (mean ± σ):      2.322 s ±  0.042 s    [User: 25.047 s, System: 4.994 s]
After:
  Time (mean ± σ):      2.269 s ±  0.023 s    [User: 24.264 s, System: 5.174 s]

1.02 ± 0.02 times faster than debuginfod-rs-before
```